+++ .rubocop_todo.yml
@@ -37,13 +37,13 @@ Naming/HeredocDelimiterCase:
 
 Naming/HeredocDelimiterNaming:
   Exclude:
-    - 'app/presenters/hyrax/file_set_presenter.rb'
+    - 'app/presenters/concerns/embed_code_presenter.rb'
     - 'lib/spec/e_pub/bridge_to_webgl_spec.rb'
     - 'lib/spec/e_pub/cfi_spec.rb'
     - 'lib/spec/e_pub/chapter_spec.rb'
     - 'lib/spec/e_pub/chapter_presenter_spec.rb'
     - 'lib/spec/e_pub/snippet_spec.rb'
-    - 'spec/presenters/hyrax/file_set_presenter_spec.rb'
+    - 'spec/presenters/concerns/embed_code_presenter_spec.rb'
 
 Naming/MemoizedInstanceVariableName:
   Exclude:
