+++ app/views/hyrax/file_sets/_show_actions.html.erb
@@ -6,7 +6,7 @@
       <%= "Back to #{parent.human_readable_type}" %>
     </a>
   <% end %>
-  <% if can?(:edit, presenter) || presenter.allow_embed? %>
+  <% if presenter.embeddable_type? && (can?(:edit, presenter) || presenter.allow_embed?) %>
     <form action="#" onsubmit="CopyToClipboard()" style="display:inline;">
       <textarea  id="embedcode" style="display:none;" readonly="readonly"><%= presenter.embed_code %></textarea>
       <button type="submit" class="btn btn-default btn-lg">&lt;html /&gt;</button>
